Refactored some of the component names.
