+++ src/router.ts
@@ -1,6 +1,6 @@
 import Vue from 'vue';
 import Router from 'vue-router';
-import Home from './views/Home.vue';
+import HomePage from './pages/HomePage.vue';
 
 Vue.use(Router);
 
@@ -12,7 +12,7 @@ export default new Router({
           path: '/',
           alias: '/home',
           name: 'home',
-          component: Home,
+          component: HomePage,
       },
   ],
 });
